Renamed workflow to deadline check
